+++ lib/predicator.ex
@@ -651,7 +651,7 @@ defmodule Predicator do
   ## Examples
 
       iex> Predicator.isa_version()
-      4
+      5
   """
   @spec isa_version() :: pos_integer()
   defdelegate isa_version(), to: Instructions